backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
celery==5.3.4
redis==5.0.1
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9

